+++ src/main/java/service/test/OwnerService.java
@@ -7,12 +7,12 @@ import java.util.List;
 
 public interface OwnerService {
     List<Owner> getOwners();
-    Owner getById(int id);
+    Owner getById(long id);
     List<Owner> create(Owner owner);
-    List<Owner> delete(int id);
+    List<Owner> delete(long id);
     List<Owner> search(String keyword, String sort);
-    List<Owner> edit(int id, Owner owner);
-    boolean checkId(int id);
+    List<Owner> edit(long id, Owner owner);
+    boolean checkId(long id);
     List<OwnerDTO> getDTO(List<Owner> owners);
 
 }
